requirements.txt
streamlit as st
dotenv 
langchain_groq 
langchain.memory 
langchain.chains 
langchain_core.prompts